moved generateImpressions into testUtils
