+++ src/impressions/observer/__tests__/observer.spec.js
@@ -1,20 +1,7 @@
 import tape from 'tape-catch';
 import ImpressionObserver from '../observer';
 import hasher from '../../hasher';
-
-const generateImpressions = count => {
-  const impressions = [];
-  for (let i = 0; i < count; i++) {
-    impressions.push({
-      keyName: `key_${i}`,
-      feature: `feature_${i % 10}`,
-      label: (i % 2 === 0) ? 'in segment all' : 'whitelisted',
-      changeNumber: i * i,
-      time: Date.now()
-    });
-  }
-  return impressions;
-};
+import { generateImpressions } from './testUtils';
 
 tape('Murmur 128 / Impression Observer Basic Functionality', assert => {
   const observer = new ImpressionObserver(5, hasher.hashImpression128);
